move ConstantFun to akka.util
