+++ akka-actor/src/main/scala/akka/util/ConstantFun.scala
@@ -1,7 +1,7 @@
 /**
  * Copyright (C) 2015-2017 Lightbend Inc. <http://www.lightbend.com>
  */
-package akka.stream.impl
+package akka.util
 
 import akka.annotation.InternalApi
 import akka.japi.function.{ Function ⇒ JFun, Function2 ⇒ JFun2 }
@@ -28,8 +28,7 @@ import akka.japi.{ Pair ⇒ JPair }
   def scalaAnyToNone[A, B]: A ⇒ Option[B] = none
   def scalaAnyTwoToNone[A, B, C]: (A, B) ⇒ Option[C] = two2none
   def javaAnyToNone[A, B]: A ⇒ Option[B] = none
-
-  val conforms = (a: Any) ⇒ a
+  def nullFun[T] = _nullFun.asInstanceOf[Any ⇒ T]
 
   val zeroLong = (_: Any) ⇒ 0L
 
@@ -37,8 +36,12 @@ import akka.japi.{ Pair ⇒ JPair }
 
   val oneInt = (_: Any) ⇒ 1
 
-  val none = (_: Any) ⇒ None
+  private val _nullFun = (_: Any) ⇒ null
 
-  val two2none = (_: Any, _: Any) ⇒ None
+  private val conforms = (a: Any) ⇒ a
+
+  private val none = (_: Any) ⇒ None
+
+  private val two2none = (_: Any, _: Any) ⇒ None
 
 }
