+++ akka-stream-tests-tck/src/test/scala/akka/stream/tck/FlattenTest.scala
@@ -3,9 +3,9 @@
  */
 package akka.stream.tck
 
-import akka.stream.impl.ConstantFun
 import akka.stream.scaladsl.Sink
 import akka.stream.scaladsl.Source
+import akka.util.ConstantFun
 import org.reactivestreams.Publisher
 
 class FlattenTest extends AkkaPublisherVerification[Int] {
